minor fix for the data import
